zigbee2mqtt: Add explicit dependency on mosquitto.service
This should fix race condition when mosquitto is stopped already while
z2m is still trying to stop. This race causes z2m to error out and
delays the service for few minutes until it reaches timeout and gets
killed. I run mosquitto and z2m on the same host, it makes sense to have
this dependency.